Reformat src files 	- Fix all warnings

+++ src/cbs/SingleAgentSolver.cpp
@@ -1,13 +1,11 @@
 #include "SingleAgentSolver.h"
 
-
 list<int> SingleAgentSolver::getNextLocations(int curr) const // including itself and its neighbors
 {
 	list<int> rst = instance.getNeighbors(curr);
 	rst.emplace_back(curr);
 	return rst;
 }
-
 
 void SingleAgentSolver::compute_heuristics()
 {
@@ -23,44 +21,44 @@ void SingleAgentSolver::compute_heuristics()
 		struct compare_node
 		{
 			// returns true if n1 > n2 (note -- this gives us *min*-heap).
-			bool operator()(const Node& n1, const Node& n2) const
+			bool operator()(const Node &n1, const Node &n2) const
 			{
 				return n1.value >= n2.value;
 			}
-		};  // used by OPEN (heap) to compare nodes (top of the heap has min f-val, and then highest g-val)
+		}; // used by OPEN (heap) to compare nodes (top of the heap has min f-val, and then highest g-val)
 	};
 
-  my_heuristic.resize(goal_location.size());
-  heuristic_landmark.resize(goal_location.size(), 0);
+	my_heuristic.resize(goal_location.size());
+	heuristic_landmark.resize(goal_location.size(), 0);
 
-  for (int i = 0; i < goal_location.size(); i++){
-    my_heuristic[i].resize(instance.map_size, MAX_TIMESTEP);
+	for (int i = 0; i < (int)goal_location.size(); i++)
+	{
+		my_heuristic[i].resize(instance.map_size, MAX_TIMESTEP);
 
-    // generate a heap that can save nodes (and a open_handle)
-    boost::heap::pairing_heap<Node, boost::heap::compare<Node::compare_node>> heap;
+		// generate a heap that can save nodes (and a open_handle)
+		boost::heap::pairing_heap<Node, boost::heap::compare<Node::compare_node>> heap;
 
-    Node root(goal_location[i], 0);
-    my_heuristic[i][goal_location[i]] = 0;
-    heap.push(root);  // add root to heap
-    while (!heap.empty())
-      {
-        Node curr = heap.top();
-        heap.pop();
-        for (int next_location : instance.getNeighbors(curr.location))
-          {
-            if (my_heuristic[i][next_location] > curr.value + 1)
-              {
-                my_heuristic[i][next_location] = curr.value + 1;
-                Node next(next_location, curr.value + 1);
-                heap.push(next);
-              }
-          }
-      }
-  }
+		Node root(goal_location[i], 0);
+		my_heuristic[i][goal_location[i]] = 0;
+		heap.push(root); // add root to heap
+		while (!heap.empty())
+		{
+			Node curr = heap.top();
+			heap.pop();
+			for (int next_location : instance.getNeighbors(curr.location))
+			{
+				if (my_heuristic[i][next_location] > curr.value + 1)
+				{
+					my_heuristic[i][next_location] = curr.value + 1;
+					Node next(next_location, curr.value + 1);
+					heap.push(next);
+				}
+			}
+		}
+	}
 
-
-  for (int i = goal_location.size() - 2; i >= 0; i--){
-    heuristic_landmark[i] = heuristic_landmark[i + 1] + my_heuristic[i + 1][goal_location[i]];
-  }
-
+	for (int i = goal_location.size() - 2; i >= 0; i--)
+	{
+		heuristic_landmark[i] = heuristic_landmark[i + 1] + my_heuristic[i + 1][goal_location[i]];
+	}
 }
